make project more clear

+++ source/panel/template/inc/JsonHelper.php
@@ -2,16 +2,44 @@
 
 class JsonHelper
 {
+    private static $rootPath;
+    private static $path = "";
+
+    private static function  getSiteJson()
+    {
+        JsonHelper::$rootPath = $_SERVER['DOCUMENT_ROOT'] . "/PRCMS_Panel/";
+        JsonHelper::$path = JsonHelper::$rootPath . "main.json";
+
+        $projectsAddress = 'projects/';
+        $realPath = JsonHelper::$rootPath . $projectsAddress;
+//        $logFile = fopen(JsonHelper::$rootPath . "log.txt", "w+") or die("Unable to open file!");
+        $log = "";
+//        $path = JsonHelper::$rootPath . "main.json";
+        if (!file_exists(JsonHelper::$path)) {
+//            $log .= ("There is no any project json file.");
+//            fwrite($logFile, $log);
+//            fclose($logFile);
+            die("cannot find json file.");
+        } else {
+            $string = file_get_contents(JsonHelper::$path);
+            $siteJson = json_decode($string, true);
+
+        }
+        return $siteJson;
+    }
+
     static function updateSiteJson($pName, $pLevel, $pPictures, $pDetail, $pThumbnail, $pCoordX, $pCoordY)
     {
-        $pictureAddresses = explode(";", $pPictures);
         $rootPath = $_SERVER['DOCUMENT_ROOT'] . "/PRCMS_Panel/";
+//        $log .= "root path is $rootPath";
         $projectsAddress = 'projects/';
         $realPath = $rootPath . $projectsAddress;
+
         $levelPath = $realPath . $pLevel . "/";
         $projectsPath = $levelPath . $pName . "/";
         JsonHelper::make_path($projectsPath);
 
+        $pictureAddresses = explode(";", $pPictures);
         $projectJson = JsonHelper::makeJsonProject($pictureAddresses, $projectsPath, $pName, $pDetail, $pThumbnail, $pCoordX, $pCoordY);
 
         $carousel = array();
@@ -30,7 +58,7 @@ class JsonHelper
         }
 
 
-        $l_index = getLevelIndex($siteJson, $pLevel);
+        $l_index = JsonHelper::getLevelIndex($siteJson, $pLevel);
         if ($l_index < 0) {
             $levelJson = array("caption" => $pLevel, "projects" => array());
             $l_index = array_push($siteJson["projectSections"], $levelJson) - 1;
@@ -53,6 +81,59 @@ class JsonHelper
         fwrite($siteFile, json_encode($siteJson));
         fclose($siteFile);
 
+    }
+
+    static function deleteProject($pName, $pLevel)
+    {
+        $siteJson = JsonHelper::getSiteJson();
+
+        $l_index = JsonHelper::getLevelIndex($siteJson, $pLevel);
+//        $log .= "level index = $l_index\n";
+        if ($l_index < 0) {
+//            $log .= ("There is no such level.");
+//            fwrite($logFile, $log);
+//            fclose($logFile);
+            return;
+        }
+        $levelJson = $siteJson["projectSections"][$l_index];
+        $p_index = JsonHelper::getProjectIndex($levelJson, $pName);
+//        $log .= "project index = $p_index\n";
+        if ($p_index < 0) {
+//            $log .= ("There is no such project.");
+//            fwrite($logFile, $log);
+//            fclose($logFile);
+            return;
+        }
+        unset($levelJson["projects"][$p_index]);
+
+//unset($siteJson["projectSections"][$l_index]["projects"][$p_index]);
+        $siteJson["projectSections"][$l_index] = $levelJson;
+        JsonHelper::writeJson(JsonHelper::$path, $siteJson);
+
+
+//        $log .= "Successful\n";
+//        fwrite($logFile, $log);
+//        fclose($logFile);
+    }
+
+    static function writeJson($filePath, $json)
+    {
+        $siteFile = fopen($filePath, "w+") or die("Unable to open file!");
+        fwrite($siteFile, json_encode($json));
+        fclose($siteFile);
+    }
+
+    static function getProjectIndex($levelJson, $projectName)
+    {
+        $i = 0;
+        foreach ($levelJson["projects"] as $project) {
+
+            if ($project["name"] === $projectName) {
+                return $i;
+            }
+            $i++;
+        }
+        return -1;
     }
 
     static function findFileName($path)
@@ -126,10 +207,14 @@ class JsonHelper
         $gallery = array();
         $i = 0;
         foreach ($pictureAddresses as $pictureAddress) {
+            if ($pictureAddress === "") {
+                continue;
+            }
             $fileName = JsonHelper::findFileName($pictureAddress);
             $newRealPath = $projectsPath . $fileName;
 //    $newPath = $imagesPath . $fileName;
 
+            echo "Picture address: $pictureAddress , newRealPath : $newRealPath";
             if (copy($pictureAddress, $newRealPath)) {
                 unlink($pictureAddress);
                 $picture = array();
